+++ package.json
@@ -31,7 +31,7 @@
     "@kong-ui-public/entities-gateway-services": "^2.2.9",
     "@kong-ui-public/entities-key-sets": "^2.1.13",
     "@kong-ui-public/entities-keys": "^2.1.13",
-    "@kong-ui-public/entities-plugins": "^2.4.11-pr.1004.ddca5d18.0",
+    "@kong-ui-public/entities-plugins": "^2.4.12",
     "@kong-ui-public/entities-routes": "^2.2.9",
     "@kong-ui-public/entities-shared": "^2.2.0",
     "@kong-ui-public/entities-snis": "^2.1.13",
